update cleaning if not ran in a week

+++ bin/CleanUp.py
@@ -48,6 +48,7 @@ def CleanUpLogs(path):
 
         os.system("ls -l " + logspace1 + " > " + logspace1 + "/file_dates.txt")
         for line2 in open(logspace1 + "/file_dates.txt", 'r'):
+            is_deleted=False
             if not ".txt" in line2:
 
                 entries = line2.split()
@@ -64,6 +65,7 @@ def CleanUpLogs(path):
                                 
                         if n_previous_jobs == 0:
                             os.system(" rm -r " + logspace1 + "/" + entries[8])
+                            is_deleted=True
                             print "Deleting directory "  + logspace1 + "/" + entries[8] +" since this is made on " + os.getenv("HOSTNAME") + " but no jobs running on this machine."
                           
                     nfiles=0
@@ -86,12 +88,16 @@ def CleanUpLogs(path):
                         days_to_keep=1
                     if nfiles != 0 :
                         days_to_keep=2
-                        
+      
+
+                     
                     if month_file not in month:
                         if int(date) > int(days_to_keep):
-                            print "Log file older than one week: Removing " + logspace1 + entries[8]
-                            os.system(" rm -r " + logspace1 + "/" + entries[8])
+                            if not is_deleted:
+                                print "Log file older than one week: Removing " + logspace1 + entries[8]
+                                os.system(" rm -r " + logspace1 + "/" + entries[8])
                     elif int(date_file) < (int(date)-int(days_to_keep)):
-                        print "Log file older than one week: Removing " + logspace1 + entries[8]   
-                        os.system(" rm -r " + logspace1 + "/" + entries[8])        
+                        if not is_deleted:
+                            print "Log file older than one week: Removing " + logspace1 + entries[8]   
+                            os.system(" rm -r " + logspace1 + "/" + entries[8])
